Cell.view -> fragment
View is also used by Morepath, but here it has a different meaning.
Fragment also describes the intended use better: most of the time fragments are used to
factor out parts from a template that becomes too big or are meant to be
reused. They can be used as an alternative to Jinja macros.