Implement CellApp
A CellApp is a Morepath app that knows something about Cells.
Cells can be registered as view models for a specific model class and a
view name (optional). The @App.cell decorator is an alternative to the
model attribute of Cell subclasses.
The methods of Cell that return other cells support both styles for now.