Update deps, Python 3.9 and Jinja 3, remove app template dir
The reason for the package loading error in ekklesia-portal was a
missing global templates dir. The error message is confusing because
it only mentions the package name and not the missing template dir itself.

This removes support for the global app template dir altogether.
Using concepts is better.

Our custom JinjaCellContext needed an update because the upstream method
signature changed.