Log tween: wrap unhandled request exceptions, SQL logging
If a request raises an exception that is not a HTTPError (these are
handled by Morepath), wrap it in a UnhandledRequestError and add
the current Eliot task UUID and a XID which is a string that can
be used to identify an exception easily in logs. XID can be shown
to the user to give them a way to refer to an error they encountered.

SQL statement history can be enabled with database.enable_statement_history.
This doesn't enable output. Enable database.print_sql_statements
to print all SQL after a request.