Add init_logging with an exception extractor, better_exceptions
Based on the code from ekklesia-portal. Adds a custom exception
extractor for Eliot logging that supplies additional fields for
exceptions: a traceback and data passed to the exception via arguments.
Exception data can be overridden by adding a __structlog__ method
to the exception. The idea is from the structlog package, though
we don't use structlog here.
Cause and context from the exception are rendered separately if provided
instead of including them in the traceback as traceback.format_exception
does.

The extractor has a special case for UnhandledRequestException
because they are always caused by another exception and have a
xid which uniquely identifies them. The XID can be used to look up
the log entry.

Tracebacks are rendered with better_exceptions if BETTER_EXCEPTIONS
is set at startup.