logging: make output stream configurable, use stdout as default
We used stderr for logging before but it's too annoying because
plain messages on stderr (from gunicorn, for example) mix with the
JSON log output. Outputting the JSON on stdout is nicer for piping
it to a formatter like eliot-tree.

The init_logging() function now returns when there's already a
handler on the root logger to avoid setting up logging twice.
This way, scripts can call set up logging in a custom way before
Ekklesia app stuff is imported which may set up logging on import.